factor out TxFactoryBase to its dedicated files
