+++ src/blsct/wallet/txfactory_base.h
@@ -1,0 +1,41 @@
+// Copyright (c) 2024 The Navio developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#include <blsct/arith/mcl/mcl.h>
+#include <blsct/wallet/address.h>
+#include <blsct/wallet/txfactory_global.h>
+#include <primitives/transaction.h>
+
+namespace blsct {
+
+struct InputCandidates {
+    CAmount amount;
+    MclScalar gamma;
+    blsct::PrivateKey spendingKey;
+    TokenId token_id;
+    COutPoint outpoint;
+    bool is_staked_commitment;
+};
+
+class TxFactoryBase
+{
+protected:
+    CMutableTransaction tx;
+    std::map<TokenId, std::vector<UnsignedOutput>> vOutputs;
+    std::map<TokenId, std::vector<UnsignedInput>> vInputs;
+    std::map<TokenId, Amounts> nAmounts;
+
+public:
+    TxFactoryBase(){};
+
+    void AddOutput(const SubAddress& destination, const CAmount& nAmount, std::string sMemo, const TokenId& token_id = TokenId(), const CreateTransactionType& type = NORMAL, const CAmount& minStake = 0, const bool& fSubtractFeeFromAmount = false);
+
+    bool AddInput(const CAmount& amount, const MclScalar& gamma, const blsct::PrivateKey& spendingKey, const TokenId& token_id, const COutPoint& outpoint, const bool& rbf = false);
+
+    std::optional<CMutableTransaction> BuildTx(const blsct::DoublePublicKey& changeDestination, const CAmount& minStake = 0, const CreateTransactionType& type = NORMAL, const bool& fSubtractedFee = false);
+
+    static std::optional<CMutableTransaction> CreateTransaction(const std::vector<InputCandidates>& inputCandidates, const blsct::DoublePublicKey& changeDestination, const SubAddress& destination, const CAmount& nAmount, std::string sMemo, const TokenId& token_id = TokenId(), const CreateTransactionType& type = NORMAL, const CAmount& minStake = 0);
+};
+
+} // namespace blsct
